chore: impl .fund()
